Add category tab in admin layout

+++ OfficeMenStore/OfficeMenStore.Application/Models/CartItem/GetAllCartItemResponse.cs
@@ -13,9 +13,9 @@ namespace OfficeMenStore.Application.Models.CartItem
         public string ProductName { get; set; }
         public string ProductImage { get; set; }
         public string SizeName { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public int Quanntity { get; set; }
-        public int Subtotal { get; set; }
+        public decimal Subtotal { get; set; }
         public bool IsDeleted { get; set; }
      }
 }
